requirements.txt
asgiref==3.3.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.4
Django==3.1.7
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.12.2
djangorestframework-simplejwt==5.1.0
drf-yasg==1.20.0
filelock==3.6.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.1
mysqlclient==2.1.0
packaging==21.3
Pillow==9.0.1
platformdirs==2.5.1
PyJWT==2.3.0
pyparsing==3.0.7
pytz==2021.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
typing_extensions==4.1.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.13.2
